Moved DirectionalNPCComponent to npc package in components
